Update a site to Atomic
The [classic editor is deprecated](https://github.com/wordpress-mobile/WordPress-iOS/pull/16008) and is [now being removed](https://github.com/wordpress-mobile/WordPress-iOS/pull/15766) for Simple sites (except when editing existing posts with classic content or when using the iOS share extension).

This means that on Simple sites, the site settings screen will no longer have a "Enable block editor" switch. WPiOS has a UI test suite, `EditorAztecTests`, that runs on a mock Simple site ("Tri-County Real Estate") and tests various features of the classic editor.

Since the classic editor is going away for Simple sites, I think it makes sense to make `EditorAztecTests` run on an Atomic site instead of a Simple site. On WPiOS, this mock site is used for all other UI tests, so this has the side-effect of switching all WPiOS UI tests to run on a mock Atomic site instead of a mock Simple site. The same could be true for WPAndroid and the same considerations might apply.