Add organization_id value to stub for WP.com me/sites
Relates to the changes in:

- https://github.com/wordpress-mobile/WordPressKit-iOS/pull/445
- https://github.com/wordpress-mobile/WordPress-iOS/pull/17154

By looking at the changes in the PRs listed above, I'm guessing the key
has been in the response from the live API for a while, but we never
updated this particular mock response because we hadn't needed the value
till now.

I noticed that the Jetpack version of me/sites does include
`organization_id`, so we can say that the value had been in the response
at least since May 2021.
See
https://github.com/wordpress-mobile/WordPressMocks/commit/caa9d47f8e3e26bfc087e8b60379c514feba77bf