Bump version from 0.0.14 to 0.0.15
Note that in this instance we are not going through a beta because today
is also code freeze for WordPress iOS and I just merged the PR, #35,
that required this new version.

Basically, I'm taking a shortcut and not releasing a beta because I
would override that beta with a stable version immediately.